Exceptions ok

+++ src/app-base/APIEmpresarial/APIEmpresarial/Services/CategoriasService.cs
@@ -9,13 +9,14 @@ namespace APIEmpresarial.Services
     public class CategoriasService : ICategoriaInterface
     {
         private readonly AppDbContext _context;
+
         public CategoriasService(AppDbContext context)
         {
             _context = context;
         }
         public void Create(Categoria categoria)
         {
-            if (categoria != null)
+            if(_context.Categorias is not null)
             {
                 _context.Categorias.Add(categoria);
                 _context.SaveChanges();
@@ -23,25 +24,52 @@ namespace APIEmpresarial.Services
         }
         public ActionResult<IEnumerable<Categoria>> GetAll()
         {
-            return _context.Categorias.ToList();
+            if(_context.Categorias is not null)
+            {
+                return _context.Categorias.AsNoTracking().Include(p => p.Livros).Where(c => c.CategoriaId <= 5).ToList();
+            }
+            else
+            {
+                return new BadRequestResult();
+            }
         }
         public ActionResult<Categoria> GetCategoria(int id)
         {
-            var categoria = _context.Categorias.FirstOrDefault(p => p.CategoriaId == id);
-            return categoria;
+            if (_context.Categorias is not null)
+            {
+                var categoria = _context.Categorias.FirstOrDefault(p => p.CategoriaId == id);
+                if (categoria != null)
+                {
+                    return categoria;
+                }
+            }
+
+            return new NotFoundResult();
         }
         public void Delete(int id)
         {
-            var categoria = _context.Categorias.FirstOrDefault(p => p.CategoriaId == id);
-            _context.Categorias.Remove(categoria);
-            _context.SaveChanges();
+            if (_context.Categorias is not null)
+            {
+                var categoria = _context.Categorias.FirstOrDefault(p => p.CategoriaId == id);
+                if (categoria != null)
+                {
+                    _context.Categorias.Remove(categoria);
+                    _context.SaveChanges();
+                }
+            }
         }
         public void Update(int id)
         {
-            var categoria = _context.Categorias.FirstOrDefault(p => p.CategoriaId == id);
-            _context.Categorias.Entry(categoria).State = EntityState.Modified;
-            _context.SaveChanges();
+            if (_context.Categorias is not null)
+            {
+                var categoria = _context.Categorias.FirstOrDefault(p => p.CategoriaId == id);
+                if (categoria != null)
+                {
+                    _context.Categorias.Entry(categoria).State = EntityState.Modified;
+                    _context.SaveChanges();
+                }
+            }
         }
-        
     }
 }
+
